requirements.txt
streamlit
google-generativeai
python-dotenv
pypdf
polars
numpy